Clean up temporary GNUPGHOME directories.
Previously, these folders wouldn't get cleaned up. Even adding a command
`rm -rf $GNUPGHOME` at the end didn't fix it, since the script would
exit on any failures of `gpg-expiration.py`. Using `trap` is the proper
way to handle this. We don't need to worry about resetting GNUPGHOME
unless this script is getting sourced. When it's executed normally,
changes to the environment don't propagate to parent processes, so
the value of GNUPGHOME before the script is run is retained.