Don't try to guess when running `gcm`
Without this change, if there's a remote `main` branch, this will check
out a new branch to track that branch, even if the intent is to check
out the `master` branch. This only really comes in handy when a repo has
both `main` and `master` branches, but the latter is the one that is
intended.