[brew.up] Convert Brew update script to bash
This used zsh for the `-q` flag that would take a single character
response without needing the user to press "enter". Now, the user can
just hit "enter" to confirm, and since the default is "yes", it's still
only a single keypress per command to confirm. The script requires an
explicit "n*" response (or Ctrl-C of course) to abort.